factor json data structures
