+++ crates/rust-project-goals-cli-llm/src/updates.rs
@@ -1,5 +1,6 @@
 use anyhow::Context;
 use chrono::{Datelike, NaiveDate};
+use rust_project_goals::json::GithubIssueState;
 use std::collections::BTreeMap;
 use std::io::Write;
 use std::path::Path;
@@ -8,9 +9,8 @@ use rust_project_goals::util::comma;
 
 use rust_project_goals::gh::{
     issue_id::{IssueId, Repository},
-    issues::{list_issue_titles_in_milestone, ExistingGithubComment, ExistingIssueState},
+    issues::{list_issue_titles_in_milestone, ExistingGithubComment, checkboxes},
 };
-use rust_project_goals::json::checkboxes;
 use rust_project_goals::gh::issues::ExistingGithubIssue;
 use crate::llm::LargeLanguageModel;
 use crate::templates::{self, Updates, UpdatesGoal};
@@ -171,7 +171,7 @@ async fn prepare_flagship_goals(
             }
             .url(),
             progress,
-            is_closed: issue.state == ExistingIssueState::Closed,
+            is_closed: issue.state == GithubIssueState::Closed,
             updates_markdown: summary,
         });
 
@@ -246,7 +246,7 @@ async fn prepare_other_goals(
                 number: issue.number,
             }
             .url(),
-            is_closed: issue.state == ExistingIssueState::Closed,
+            is_closed: issue.state == GithubIssueState::Closed,
             updates_markdown: summary,
             progress: checkboxes(&issue),
         };
